Refact: Ajusta runTimeException para tratamento de exceções melhores e mais claros para erros de requisições

+++ src/main/java/dev/Cursos/cursos/Aulas/AulasService.java
@@ -1,13 +1,14 @@
 package dev.Cursos.cursos.Aulas;
 
-import java.util.List;
-
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import dev.Cursos.cursos.Aulas.dto.AulasRequestDTO;
 import dev.Cursos.cursos.Aulas.dto.AulasResponseDTO;
 import dev.Cursos.cursos.Modulo.ModuloModel;
 import dev.Cursos.cursos.Modulo.ModuloRepository;
+import dev.Cursos.cursos.exceptions.ResourceNotFoundException;
 
 @Service
 public class AulasService {
@@ -22,22 +23,21 @@ public class AulasService {
         this.aulasMapper = aulasMapper;
     }
 
-    public List<AulasResponseDTO> getAllAulas(){
-        return aulasRepository.findAll()
-            .stream()
-            .map(aulasMapper::toResponse)
-            .toList();
+    public Page<AulasResponseDTO> getAllAulas(Pageable pageable) {
+        return aulasRepository.findAll(pageable)
+                .map(aulasMapper::toResponse);
     }
+
 
     public AulasResponseDTO getAulasById(Long id) {
         return aulasRepository.findById(id)
         .map(aulasMapper::toResponse)
-        .orElseThrow(()-> new RuntimeException("Aula não encontrada"));
+        .orElseThrow(()-> new ResourceNotFoundException("Aula com ID " + id + " não encontrada."));
     }
 
     public AulasResponseDTO createAulas(AulasRequestDTO dto) {
         ModuloModel modulo = moduloRepository.findById(dto.id_modulo())
-                .orElseThrow(() -> new RuntimeException("Módulo não encontrado."));
+                .orElseThrow(() -> new ResourceNotFoundException("Módulo com ID " + dto.id_modulo() + " não encontrado."));
 
         AulasModel aulas = aulasMapper.toModel(dto, modulo);
         AulasModel saved = aulasRepository.save(aulas);
@@ -45,6 +45,9 @@ public class AulasService {
         return aulasMapper.toResponse(saved);
     }
     public void deleteAulas(Long id) {
+        if (!aulasRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Aula com ID " + id + " não encontrada para exclusão.");
+        }
         aulasRepository.deleteById(id);
     }
 }
