+++ src/main/java/dev/Cursos/cursos/Post/PostService.java
@@ -1,13 +1,14 @@
 package dev.Cursos.cursos.Post;
 
-import java.util.List;
-
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import dev.Cursos.cursos.Curso.CursoModel;
 import dev.Cursos.cursos.Curso.CursoRepository;
 import dev.Cursos.cursos.Post.dto.PostRequestDTO;
 import dev.Cursos.cursos.Post.dto.PostResponseDTO;
+import dev.Cursos.cursos.exceptions.ResourceNotFoundException;
 
 @Service
 public class PostService {
@@ -22,22 +23,20 @@ public class PostService {
         this.postMapper = postMapper;
     }
 
-    public List<PostResponseDTO> getAllPosts() {
-        return postRepository.findAll()
-                .stream()
-                .map(postMapper::toResponse)
-                .toList();
+    public Page<PostResponseDTO> getAllPosts(Pageable pageable) {
+        return postRepository.findAll(pageable)
+                .map(postMapper::toResponse);
     }
 
     public PostResponseDTO getPostById(Long id) {
         return postRepository.findById(id)
                 .map(postMapper::toResponse)
-                .orElseThrow(() -> new RuntimeException("Post não encontrado."));
+                .orElseThrow(() -> new ResourceNotFoundException("Post com ID " + id + " não encontrada."));
     }
 
     public PostResponseDTO createPost(PostRequestDTO dto) {
         CursoModel curso = cursoRepository.findById(dto.cursoId())
-                .orElseThrow(() -> new RuntimeException("Curso não encontrado."));
+                .orElseThrow(() -> new ResourceNotFoundException("Curso com ID " + dto.cursoId() + " não encontrado."));
 
         PostModel post = postMapper.toModel(dto, curso);
         PostModel saved = postRepository.save(post);
@@ -46,6 +45,9 @@ public class PostService {
     }
 
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Post com ID " + id + " não encontrada para exclusão.");
+        }
         postRepository.deleteById(id);
     }
 }
